requirements.txt
xarray
netCDF4
pandas
numpy
psycopg2-binary
SQLAlchemy
tqdm
sentence-transformers
chromadb
python-dotenv
supabase
requests
aiohttp

# FastAPI dependencies
fastapi
uvicorn[standard]
python-multipart
aiofiles
